+++ akka-docs/src/test/java/jdocs/stream/operators/SourceDocExamples.java
@@ -5,11 +5,13 @@
 package jdocs.stream.operators;
 
 //#imports
+//#range-imports
 import akka.NotUsed;
 import akka.actor.ActorSystem;
 import akka.stream.ActorMaterializer;
 import akka.stream.Materializer;
 import akka.stream.javadsl.Source;
+//#range-imports
 
 import java.util.Arrays;
 
@@ -30,6 +32,31 @@ public class SourceDocExamples {
         Source<String, NotUsed> words = Source.from(Arrays.asList(text.split("\\s")));
         words.runForeach(System.out::println, materializer);
         //#source-from-example
+    }
+
+    static void rangeExample() {
+
+        final ActorSystem system = ActorSystem.create("Source");
+        final Materializer materializer = ActorMaterializer.create(system);
+
+        //#range
+
+        Source<Integer, NotUsed> source = Source.range(1, 100);
+
+        //#range
+
+        //#range
+        Source<Integer, NotUsed> sourceStepFive = Source.range(1, 100, 5);
+
+        //#range
+
+        //#range
+        Source<Integer, NotUsed> sourceStepNegative = Source.range(100, 1, -1);
+        //#range
+
+        //#run-range
+        source.runForeach(i -> System.out.println(i), materializer);
+        //#run-range
     }
 
 }
